Коррекция теста ViewModel, Coroutine
