+++ build.gradle
@@ -55,7 +55,7 @@ subprojects {
                 // Mockito
                 testImplementation 'org.junit.jupiter:junit-jupiter-api:5.5.2'
                 testImplementation "org.mockito:mockito-core:3.10.0"
-                testImplementation 'org.mockito:mockito-inline:2.8.9'
+                testImplementation 'org.mockito:mockito-inline:2.13.0'
                 testImplementation('com.nhaarman:mockito-kotlin:1.5.0') {
                     exclude group: 'org.jetbrains.kotlin'
                     exclude group: 'org.mockito'
@@ -86,9 +86,7 @@ subprojects {
 
                 testImplementation 'androidx.arch.core:core-testing:2.2.0'
                 implementation 'com.jakewharton.retrofit:retrofit2-kotlin-coroutines-adapter:0.9.2'
-                testImplementation 'org.jetbrains.kotlinx:kotlinx-coroutines-test:1.4.3'
-
-
+                testImplementation 'org.jetbrains.kotlinx:kotlinx-coroutines-test:1.6.4'
 
             }
         }
